Support i32 for table, title for typography
